Update the parsing to allow for floats or ints
The parser has been updated to allow for either floats or ints, so that I
could set up the modulo operator for ints. In order to allow this, the
eval function is now returning a struct that holds the value(s) and a type
string. There is probably a much better way to handle a lot of this code,
but I am way out of touch with C best practices.